add order controllers

+++ src/products/products.controller.ts
@@ -1,7 +1,17 @@
-import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  NotFoundException,
+  Param,
+  Post,
+  Put,
+} from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { Product } from 'src/db';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 import { ParseUUIDPipe } from '@nestjs/common';
 
 @Controller('products')
@@ -9,17 +19,23 @@ export class ProductsController {
   constructor(private productsService: ProductsService) {}
 
   @Get('/')
-  public getAll(): any {
+  public getAll(): Product[] {
     return this.productsService.getAll();
   }
 
   @Get('/:id')
   public getById(@Param('id', new ParseUUIDPipe()) id: Product['id']) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException(`Product with id ${id} not found`);
+    }
     return this.productsService.getById(id);
   }
 
   @Delete('/:id')
   public deleteById(@Param('id', new ParseUUIDPipe()) id: Product['id']) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException(`Product with id ${id} not found`);
+    }
     this.productsService.delete(id);
     return { message: 'Product deleted' };
   }
@@ -27,5 +43,17 @@ export class ProductsController {
   @Post('/')
   create(@Body() productData: CreateProductDTO) {
     return this.productsService.create(productData);
+  }
+
+  @Put('/:id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: Product['id'],
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException(`Product with id ${id} not found`);
+    }
+    this.productsService.update(id, productData);
+    return { message: 'Product updated' };
   }
 }
